feat: add completed task in modules folder
